Move get_all_files_of_type into helper module
