+++ dmscripts/bulk_upload_documents.py
@@ -1,4 +1,3 @@
-import os
 import re
 import csv
 
@@ -36,13 +35,6 @@ def upload_file(bucket, dry_run, file_path, framework_slug, bucket_category, sup
         print(supplier_id)
     else:
         print("[Dry-run] UPLOAD: '{}' to '{}'".format(file_path, upload_path))
-
-
-def get_all_files_of_type(local_directory, file_type):
-    for root, subfolder, files in os.walk(local_directory):
-        for filename in files:
-            if filename.endswith(file_type):
-                yield os.path.join(root, filename)
 
 
 def get_supplier_id_from_framework_file_path(path):
