+++ scripts/upload-counterpart-agreements.py
@@ -42,7 +42,7 @@ import sys
 sys.path.insert(0, '.')
 
 from dmscripts.helpers.auth_helpers import get_auth_token
-from dmscripts.bulk_upload_documents import get_all_files_of_type
+from dmscripts.helpers.file_helpers import get_all_files_of_type
 from dmscripts.upload_counterpart_agreements import upload_counterpart_file
 from dmscripts.helpers import logging_helpers
 from dmscripts.helpers.logging_helpers import logging
